# api/openapi.yaml
components:
  schemas:
    AccessToken:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          pattern: Z$
          type: string
        expires_at:
          description: "From then on it no longer authenticates, and stays listed until it is revoked; null: never expires."
          format: date-time
          pattern: Z$
          type:
            - string
            - "null"
        id:
          format: uuid
          type: string
        last_used_at:
          description: "The latest request made with it, to the hour; null: never used."
          format: date-time
          pattern: Z$
          type:
            - string
            - "null"
        name:
          maxLength: 64
          minLength: 1
          type: string
      required:
        - id
        - name
        - created_at
        - last_used_at
        - expires_at
      type: object
    AccessTokens:
      additionalProperties: false
      properties:
        items:
          description: Newest first, expired ones included.
          items:
            $ref: "#/components/schemas/AccessToken"
          type: array
      required:
        - items
      type: object
...
    CreatedAccessToken:
      additionalProperties: false
      properties:
        created_at:
          format: date-time
          pattern: Z$
          type: string
        expires_at:
          description: "null: never expires."
          format: date-time
          pattern: Z$
          type:
            - string
            - "null"
        id:
          format: uuid
          type: string
        last_used_at:
          description: "Always null: it was never used."
          format: date-time
          pattern: Z$
          type:
            - string
            - "null"
        name:
          maxLength: 64
          minLength: 1
          type: string
        token:
          description: "The secret, to send as Authorization: Bearer. Shown this once: only its hash is kept."
          pattern: ^rst_[A-Za-z0-9_-]{43}$
          type: string
      required:
        - token
        - id
        - name
        - created_at
        - last_used_at
        - expires_at
      type: object
...
    Error:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - error
      type: object
    ErrorDetail:
      additionalProperties: false
      properties:
        code:
          description: Stable and machine-readable.
          enum:
            - invalid_parameter
            - invalid_body
            - unauthorized
            - invalid_credentials
            - cross_origin
            - not_found
            - method_not_allowed
            - unsupported_media_type
            - too_many_attempts
            - currency_in_use
            - without_position_taken
            - too_many_places
            - charge_has_cost
            - api_key_refused
            - instance_key
            - too_many_vehicles
            - feature_unavailable
            - broker_refused
            - insufficient_scope
            - rate_limited
            - internal
          type: string
        message:
          description: For humans; it may change.
          minLength: 1
          type: string
      required:
        - code
        - message
      type: object
...
    NewAccessToken:
      additionalProperties: false
      properties:
        expiry:
          description: How long it authenticates, from now.
          enum:
            - 30d
            - 90d
            - 365d
            - never
          type: string
        name:
          description: To tell the tokens apart, such as the program that uses it.
          examples:
            - Home Assistant
          maxLength: 64
          minLength: 1
          type: string
      required:
        - name
        - expiry
      type: object
...
  securitySchemes:
    accessToken:
      description: "A personal access token, `rst_` then 43 characters, issued by `POST /tokens`: it reads only."
      scheme: bearer
      type: http
    session:
      description: The session cookie set by `POST /session`, when the instance is served over plain http.
      in: cookie
      name: runsten_session
      type: apiKey
    sessionSecure:
      description: "The same cookie when the instance is served over https (its redirect URI is https): `Secure`, and the `__Host-` prefix binds it to the host."
      in: cookie
      name: __Host-runsten_session
      type: apiKey
info:
  contact:
    name: Runsten
    url: https://github.com/runsten-app/runsten
  description: |-
    The session, the account's vehicles, their current state (Verdandi), their
    trips and charges (Urd) with the cost of each charge, their statistics, the account's
    settings (its currency, and its places with their tariffs), and the costs entered for
    charges.

    Conventions:

    - Units are in the field names: `_km`, `_m` (metres), `_kwh`, `_pct` (state of charge, in %), `_w`, `_kw`, `_kwh_per_100km`, `_s` (seconds), `_minor` (minor units of the account's currency, as cents: an integer), `price_per_kwh` (major units of the account's currency per kWh).
      Positions are WGS 84, in degrees.
    - Times are RFC 3339, in UTC.
    - An unknown value is `null`, never zero. Every key is always present.
    - Lists are `{"items": [...]}`, never `null`. Trips and charges are paged with an
      opaque cursor, newest first.
    - Errors are `{"error": {"code", "message"}}`: the code is stable, the message is
      for humans and may change. A path under `/api/` that serves no route answers 404
      `not_found`, and a method that a path does not accept answers 405
      `method_not_allowed` with an `Allow` header.
    - Every response carries `Cache-Control: no-store`, `X-Content-Type-Options: nosniff`
      and `Referrer-Policy: no-referrer`.

    Every route but `POST /session` requires the session cookie it sets, or, to read, a
...
  license:
...
    name: AGPL-3.0-or-later
...
openapi: 3.1.0
paths:
...
  /tokens:
    get:
      operationId: listAccessTokens
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokens"
          description: The tokens, without their secrets.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`insufficient_scope`: this operation requires a session, never an access token."
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
      summary: The user's access tokens
      tags:
        - tokens
    post:
      description: Its secret is in this response only.
      operationId: createAccessToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewAccessToken"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatedAccessToken"
          description: The token, with its secret.
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`invalid_body`: not a single JSON object of the schema, a value out of its bounds, an unknown field, or more than 1 MiB; the message names the field, never its value."
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`cross_origin`: a request that changes state, from another origin. `insufficient_scope`: this operation requires a session, never an access token."
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unsupported_media_type`: the body is not `application/json`."
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
      summary: Issue an access token
      tags:
        - tokens
  /tokens/{token}:
    delete:
      description: It no longer authenticates, at once.
      operationId: deleteAccessToken
      parameters:
        - description: The token's ID.
          in: path
          name: token
          required: true
          schema:
            description: The token's ID.
            type: string
      responses:
        "204":
          description: Revoked.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`unauthorized`: no valid session, or an invalid or expired access token (with a `WWW-Authenticate` header)."
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`cross_origin`: a request that changes state, from another origin. `insufficient_scope`: this operation requires a session, never an access token."
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`not_found`: no such token of the user. Another user's token is not found either."
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: "`internal`: an internal failure, without detail; the cause is logged."
      security:
        - session: []
        - sessionSecure: []
      summary: Revoke an access token
      tags:
        - tokens
...
servers:
...